+++ core/internal/routes.js
@@ -19,9 +19,6 @@ module.exports = {
     require('./explorer/stream'),
     require('./explorer/stream-classifications')
   ],
-  guardians: [
-    require('./guardians')
-  ],
   ingest: [
     require('./ingest/stream')
   ],