+++ noncore/_cli/migrations/20220415000001-delete-filter-presets-table.js
@@ -1,0 +1,11 @@
+'use strict'
+
+module.exports = {
+  up: function (queryInterface, Sequelize) {
+    return queryInterface.sequelize.query('DROP TABLE IF EXISTS FilterPresets;')
+  },
+
+  down: function () {
+    return true // no way back!
+  }
+}