+++ noncore/_services/mail/mail-service.js
@@ -1,23 +1,8 @@
 const Converter = require('../../../common/converter')
-const Promise = require('bluebird')
 const mailing = require('./mailchimp-wrapper')
 const fs = require('fs')
 const path = require('path')
 const handlebars = require('handlebars')
-
-function sendTextMail (serviceRequest) {
-  const params = {}
-  serviceRequest = new Converter(serviceRequest, params)
-  serviceRequest.convert('email_address').toString()
-  serviceRequest.convert('recipient_name').optional().default('').toString()
-  serviceRequest.convert('subject').toString()
-  serviceRequest.convert('message').toString()
-
-  return serviceRequest.validate()
-    .then(() => {
-      return mailing.sendMail(params.email_address, params.recipient_name, params.subject, params.message)
-    })
-}
 
 function sendMessage (serviceRequest) {
   const params = {}
@@ -44,12 +29,12 @@ function sendEmail (serviceRequest) {
   serviceRequest.convert('text').optional().default('').toString()
   serviceRequest.convert('html').optional().default('').toString()
   serviceRequest.convert('subject').optional().toString()
-  serviceRequest.convert('from_email').optional().toString()
-  serviceRequest.convert('from_name').optional().toString()
-  serviceRequest.convert('merge_language').optional().toString()
+  serviceRequest.convert('from_email').toString().default('noreply@rfcx.org')
+  serviceRequest.convert('from_name').toString().default('Rainforest Connection')
+  serviceRequest.convert('merge_language').toString().default('handlebars')
   serviceRequest.convert('to').toArray()
   serviceRequest.convert('global_merge_vars').optional().toArray()
-  serviceRequest.convert('merge_vars').optional().toArray()
+  serviceRequest.convert('merge_vars').toArray().default([])
   serviceRequest.convert('important').optional().toBoolean()
   serviceRequest.convert('bcc_address').optional().toString()
 
@@ -59,10 +44,10 @@ function sendEmail (serviceRequest) {
     })
 }
 
-function renderContactFormEmail (opts) {
+function renderTemplate (tmplPath, opts) {
   return new Promise((resolve, reject) => {
     try {
-      const source = fs.readFileSync(path.join(__dirname, '../../views/email/contact-form.handlebars'), 'utf8')
+      const source = fs.readFileSync(path.join(__dirname, tmplPath), 'utf8')
       const template = handlebars.compile(source)
       resolve(template(opts))
     } catch (e) {
@@ -71,48 +56,17 @@ function renderContactFormEmail (opts) {
   })
 }
 
-function getDetectionAlertHtml () {
-  return new Promise((resolve, reject) => {
-    try {
-      const source = fs.readFileSync(path.join(__dirname, '../../views/email/detection-alert.handlebars'), 'utf8')
-      resolve(source)
-    } catch (e) {
-      reject(e)
-    }
-  })
+function renderContactFormEmail (opts) {
+  return renderTemplate('../../views/email/contact-form.handlebars', opts)
 }
 
-function getEventAlertHtml () {
-  return new Promise((resolve, reject) => {
-    try {
-      const source = fs.readFileSync(path.join(__dirname, '../../views/email/event-alert.handlebars'), 'utf8')
-      resolve(source)
-    } catch (e) {
-      reject(e)
-    }
-  })
-}
-
-function registerToAppWaitingList (serviceRequest) {
-  const params = {}
-  serviceRequest = new Converter(serviceRequest, params)
-
-  serviceRequest.convert('email_address').toString()
-  serviceRequest.convert('requested_os').optional().default('iOS').toString()
-
-  return serviceRequest.validate().then(() => {
-    return mailing.subsribeToList(process.env.MAILCHIMP_APP_WAITING_LIST, params.email_address)
-  }).then(() => {
-    return { email_address: params.email_address, method: 'subscribe_to_list', success: true }
-  })
+function getEventAlertHtml (opts) {
+  return renderTemplate('../../views/email/event-alert.handlebars', opts)
 }
 
 module.exports = {
-  registerToAppWaitingList,
-  sendTextMail,
   sendMessage,
   sendEmail,
   renderContactFormEmail,
-  getDetectionAlertHtml,
   getEventAlertHtml
 }