+++ noncore/v2/guardians/guardians-segments.js
@@ -1,7 +1,6 @@
 const express = require('express')
 const router = express.Router()
-const { httpErrorResponse } = require('../../../common/error-handling/http')
-const { ValidationError } = require('sequelize')
+const { httpErrorHandler } = require('../../../common/error-handling/http')
 const passport = require('passport')
 
 const msgSegUtils = require('../../_utils/rfcx-guardian/guardian-msg-parsing-utils').guardianMsgParsingUtils
@@ -23,17 +22,7 @@ router.route('/segments/sms')
           res.writeHead(200, { 'Content-Type': 'text/xml' })
           res.end()
         })
-        .catch(ValidationError, e => {
-          let message = 'Validation error'
-          try {
-            message = e.errors && e.errors.length ? e.errors.map((er) => er.message).join('; ') : e.message
-          } catch (err) { }
-          httpErrorResponse(req, res, 400, null, message)
-        })
-        .catch(function (err) {
-          console.error(err)
-          res.status(500).json({ message: err.message, error: { status: 500 } })
-        })
+        .catch(httpErrorHandler(req, res, 'Failed processing segment'))
     } else {
       res.writeHead(401, { 'Content-Type': 'text/xml' })
       res.end()
@@ -52,17 +41,7 @@ router.route('/segments/sbd')
           res.writeHead(200, { 'Content-Type': 'text/xml' })
           res.end()
         })
-        .catch(ValidationError, e => {
-          let message = 'Validation error'
-          try {
-            message = e.errors && e.errors.length ? e.errors.map((er) => er.message).join('; ') : e.message
-          } catch (err) { }
-          httpErrorResponse(req, res, 400, null, message)
-        })
-        .catch(function (err) {
-          console.error(err)
-          res.status(500).json({ message: err.message, error: { status: 500 } })
-        })
+        .catch(httpErrorHandler(req, res, 'Failed processing segment'))
     } else {
       res.writeHead(401, { 'Content-Type': 'text/xml' })
       res.end()