+++ tasks/event-names.js
@@ -1,5 +1,0 @@
-const SEGMENT_CREATED = 'segmentCreated'
-const EVENT_CREATED = 'event-created'
-const EVENT_UPDATED = 'event-updated'
-
-module.exports = { SEGMENT_CREATED, EVENT_CREATED, EVENT_UPDATED }