+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-## 1.1.3 (2022-04-XX)
+## 1.1.4 (2022-04-XX)
+
+### Features
+* **noncore**: Add POST /users endpoint
+* **noncore**: Endpoint for getting guardians by project id
+
+### Performance improvements
+* **noncore**: Delete code and table related to FilterPresets, Reports with Attachments, UserLocations, UserTokens, ResetPasswordTokens
+* **noncore**: Delete and stub endpoints for /v1/sites and /v1/guardians
+* **core**: Replace db query with existing attributes mapping for ingest endpoint
+
+### Bug Fixes
+* **core**: Set arbimon ingestion request timeout to 10 seconds
+
+## 1.1.3 (2022-04-05)
 
 ### Bug Fixes
 * **mqtt** Refactor GuardianMetaAssetExchangeLogs clearance by making it in one query